# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Product:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
    type: object
...
  dto.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        minimum: 0
        type: integer
    required:
    - description
    - name
    - price
    - sku
    - stock
    type: object
...
  dto.UpdateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
    type: object
  pkg.ErrorResponseSwag:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  pkg.PaginatedSuccessResponseSwag:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/pkg.PaginationMeta'
      status:
        type: string
    type: object
  pkg.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  pkg.SuccessResponseSwag:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API cho hệ thống e-commerce
  termsOfService: http://example.com/terms/
  title: E-commerce API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of products with optional filtering by
        name and SKU
      parameters:
      - description: Filter by product name
        in: query
        name: name
        type: string
      - description: Filter by product SKU
        in: query
        name: sku
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.PaginatedSuccessResponseSwag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product with the provided details
      parameters:
      - description: Create product request
        in: body
        name: createProductRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pkg.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing product by ID
      parameters:
      - description: Product ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.SuccessResponseSwag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      summary: Delete product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update the details of an existing product by ID
      parameters:
      - description: Product ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Update product request
        in: body
        name: updateProductRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/domain.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      summary: Update product
      tags:
      - Products
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"